I adjusted the width of the sidepanel columns so that input boxes would be longer.
The "" growth model option wasn't showing up in the gui, so I changed it from "" to "None"
throughout the code (there could be spots that I missed though)

In GetAgeAssign() I made the list "results" part of the returned list. I need to
think some more about the best way to construct the output of that function.

I added the "More Plots" tab (just a place holder name). For now it contains
estimated catches by brood year and age. I was thinking that these plots could
help when the glm model breaks, but unfortunately when the model breaks, the app breaks
and nothing gets displaye. We could use a try-catch on the glm model so that other
plots are produced even when the glm model fails.

The "More Plots" plots are produced by the function plotByBroodYear() at the bottom
of "functions_mitchell_model.R" (this may be doing the same thing as one
of your plot functions, sorry if that's the case I just wanted to try out adding
many plots at once). I switched from bar graph to points so it would be easier to
see all the zeros.

I'm thinking all the plots should be done by separate functions with separate calls
in the server.R file instead of within GetAgeAssign(), so the user can
pick and choose which to display. Also, this way GetAgeAssign() won't need to be
rerun every time the user changes which plots to produce (I think that would be how
it works). We could pass each function the list that gets returned by GetAgeAssign(),
e.g., see plotByBroodYear(). This will result in each function being passed more
information than it needs, but it would simplify/unify things structurally.